requirements.txt
locust
ddt~=1.1.3
pytest
pytest-parallel
pytest-xdist
pytest-html
pycco
requests
selenium
pillow
beautifulsoup4
html5lib
pytest-ordering
random-word
psycopg2-binary
Faker
